rpidistro-ffmpeg: fix compiling for arm arch with new binutils
Lately binutils (since v2.42) introduced stricter code checking
requirements, which makes an assembly label to be invalid, which
fails the build when building using Styhead or newer for 32-bit
arm arch (e.g. raspberrypi3 MACHINE)

This issue is tracked by ffmpeg at https://trac.ffmpeg.org/ticket/11074

The patch in this commit is a backport of the fix, which can be
found also in the above bugreport.